feat: unit tests for functions in the utils folder
