+++ src/common/utils/form-values/form-values.ts
@@ -6,9 +6,9 @@ export function getFormValues(elements: MalgaFormElements) {
   const formElements = getFormElements(elements)
 
   return {
-    holderName: formElements.holderName!.value,
-    cvv: formElements.cvv!.value,
-    expirationDate: formElements.expirationDate!.value,
-    number: formElements.number!.value,
+    holderName: formElements.holderName?.value,
+    cvv: formElements.cvv?.value,
+    expirationDate: formElements.expirationDate?.value,
+    number: formElements.number?.value,
   }
 }
